Disable rspec rubocop checks in integration test
The integration test uses instance variables to keep state across
different rspec examples, such as tracking the transition from unmounted
to mounted. As the state changes, so do the expectations for the next
test.

Unwinding that would take some time, time better spent converting the
test to an acceptance test. So for now disable those checks for that
test only.