(MODULES-8304) Allow spaces in mounts on Linux
This commit changes the mount type to accept whitespaces in mountpoint
names and devices on Linux. This is accomplished by replacing the
whitespace character with the ASCII code for space (\040) before writing
to /etc/fstab.

When reading mountpoints from fstab, names and devices are munged back
to the original whitespace character.

To make these changes as unintrusive as possible, the whitespace
conversion is done by implementing the `to_line` and `post_parse`
methods from the ParsedFile provider.